implementa fluxo completo de cadastro e listagem de usuários

+++ src/main/java/com/Thiago279/agendamento_medico/controller/ClientController.java
@@ -2,25 +2,15 @@ package com.Thiago279.agendamento_medico.controller;
 
 import com.Thiago279.agendamento_medico.dto.ClientRequestDTO;
 import com.Thiago279.agendamento_medico.dto.ClientUpdateDTO;
-import com.Thiago279.agendamento_medico.dto.ErroResposta;
-import com.Thiago279.agendamento_medico.dto.ScheduleUpdateDTO;
 import com.Thiago279.agendamento_medico.entity.Client;
-import com.Thiago279.agendamento_medico.entity.Schedule;
-import com.Thiago279.agendamento_medico.exceptions.ClienteNaoEncontradoException;
-import com.Thiago279.agendamento_medico.exceptions.OperacaoNaoPermitidaException;
-import com.Thiago279.agendamento_medico.exceptions.RegistroDuplicadoException;
 import com.Thiago279.agendamento_medico.service.impl.ClientService;
-import com.Thiago279.agendamento_medico.service.impl.ScheduleService;
 import io.swagger.v3.oas.annotations.Operation;
-import jakarta.persistence.EntityNotFoundException;
+import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.util.List;
-import java.util.Objects;
 
 @RestController
 @RequestMapping("clients")
@@ -47,7 +37,7 @@ public class ClientController {
 
     @Operation(summary = "Deleta cliente (id)", method = "DELETE")
     @DeleteMapping("{id}")
-    public ResponseEntity<Object> deleteClient(@PathVariable Long id ){
+    public ResponseEntity<Object> deleteClient(@PathVariable Integer id ){
         service.delete(id);
         return ResponseEntity.noContent().build();
 
@@ -55,7 +45,7 @@ public class ClientController {
 
     @Operation(summary = "atualiza dados do cliente (id)", method = "PUT")
     @PutMapping("{id}")
-    public ResponseEntity<Object> updateSchedule(@PathVariable Long id , @RequestBody @io.swagger.v3.oas.annotations.parameters.RequestBody(
+    public ResponseEntity<Object> updateSchedule(@PathVariable Integer id , @RequestBody @io.swagger.v3.oas.annotations.parameters.RequestBody(
             description = "Objeto contendo o novo telefone e/ou email do cliente",
             content = @io.swagger.v3.oas.annotations.media.Content(
                     schema = @Schema(implementation = ClientUpdateDTO.class)
