+++ src/main/java/com/Thiago279/agendamento_medico/controller/ScheduleController.java
@@ -1,15 +1,12 @@
 package com.Thiago279.agendamento_medico.controller;
 
 
-import com.Thiago279.agendamento_medico.dto.ErroResposta;
 import com.Thiago279.agendamento_medico.dto.ScheduleRequestDTO;
 import com.Thiago279.agendamento_medico.dto.ScheduleUpdateDTO;
 import com.Thiago279.agendamento_medico.entity.Schedule;
-import com.Thiago279.agendamento_medico.service.impl.ClientService;
 import com.Thiago279.agendamento_medico.service.impl.ScheduleService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import jakarta.persistence.EntityNotFoundException;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -41,14 +38,14 @@ public class ScheduleController {
 
     @Operation(summary = "Deleta agendamento de consulta", method = "DELETE")
     @DeleteMapping("{id}")
-    public ResponseEntity<Object> deleteSchedule(@PathVariable Long id ){
+    public ResponseEntity<Object> deleteSchedule(@PathVariable Integer id ){
         service.delete(id);
         return ResponseEntity.noContent().build();
     }
 
     @Operation(summary = "atualiza data de agendamento de consulta (id)", method = "PUT")
     @PutMapping("{id}")
-    public ResponseEntity<Object> updateSchedule(@PathVariable Long id , @RequestBody @Valid ScheduleUpdateDTO schedule) {
+    public ResponseEntity<Object> updateSchedule(@PathVariable Integer id , @RequestBody @Valid ScheduleUpdateDTO schedule) {
         return ResponseEntity.ok(service.update(id, schedule));
     }
 }
