+++ src/main/java/com/Thiago279/agendamento_medico/dto/ScheduleRequestDTO.java
@@ -3,7 +3,7 @@ package com.Thiago279.agendamento_medico.dto;
 import java.time.LocalDateTime;
 
 public record ScheduleRequestDTO(
-        Long clientId,
+        Integer clientId,
         LocalDateTime startDate
 ) {
